+++ src/lib/config/StringConfig.svelte
@@ -7,11 +7,13 @@
 	import Icon from '@iconify/svelte';
 	import InputField from '$lib/components/InputField.svelte';
 	import { expandedEntry } from './ExpandedEntryPopup.svelte';
+	import { getListSeparator } from '$lib/util';
 
 	export let entryId: ConfigEntryId;
 	export let isOther: boolean = false;
 
 	let content = entryId.entry.value.content as string;
+	let listSeparator = getListSeparator(entryId.entry);
 
 	function onReset(value: ConfigValue) {
 		content = value.content as string;
@@ -24,9 +26,9 @@
 </script>
 
 <div class="flex-grow relative">
-	<InputField bind:value={content} class="flex-grow" />
+	<InputField bind:value={content} />
 
-	{#if content.length > 100 || content.includes(',') || content.includes('\n')}
+	{#if content.length > 100 || content.includes('\n') || content.includes(listSeparator)}
 		<Button.Root
 			class="absolute right-1 top-1 p-1 text-slate-400 text-lg rounded-lg hover:bg-gray-800 bg-gray-900"
 			on:click={() => ($expandedEntry = entryId)}