+++ src/routes/test/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import Markdown from "$lib/components/Markdown.svelte";
+</script>
+
+<Markdown source=" <details> <summary>Click to expand</summary> ### 1.0.3 Prevent Ambigious method matching from halting patching. Prevent Transpiler IL errors from halting patching. Only log spatilization changes if anything was patched. ### 1.0.2 Fixed patching logs from classes other than Assembly-CSharp. Fixed patching not applying if only regex patches were used. Blocked trying to patch LogException. ### 1.0.1 Added versioning to the config, a warning message will appear if the config version is different than expected. Added patch for Quaternion.LookRotation, will mask the Look rotation viewing vector is zero warnings. Plugin no longer depends on Lethal Company and should work for any game. Config generation now done in separate harmony namespace and also unpatches itself. </details>" />